+++ utils.py
@@ -1,0 +1,18 @@
+import extruct
+
+def extract_metadata(text):
+    """Extract all metadata present in the page and return a dictionary of metadata lists.
+
+    Args:
+
+    Returns:
+        metadata (dict): Dictionary of json-ld, microdata, and opengraph lists.
+        Each of the lists present within the dictionary contains multiple dictionaries.
+    """
+
+    metadata = extruct.extract(text,
+                               uniform=True,
+                               syntaxes=['json-ld',
+                                         'microdata',
+                                         'opengraph'])
+    return metadata
